Refactoring, comments
