+++ exercise.main/CashRegister.cs
@@ -1,0 +1,8 @@
+﻿using exercise.main;
+
+namespace exercise.tests
+{
+    public class CashRegister 
+    {
+    }
+}
