+++ exercise.main/Discount/DiscountManager.cs
@@ -6,7 +6,6 @@ namespace exercise.main.Discount
     {
 
         List<DiscountBase> discountTypes = new List<DiscountBase>();
-        //public void addDiscountType<T>(params object[] args) where T : Discount
         private Inventory _inventory;
         public DiscountManager(Inventory inventory)
         {
@@ -14,14 +13,13 @@ namespace exercise.main.Discount
         }
         public void addDiscountType(DiscountBase discount)
         {
-            // TODO: Check for and remove identicals...
+            // Note: should check for and remove identicals...
             discountTypes.Add(discount);
         }
 
         private List<DiscountedProductCount> pickBestDeals(List<DiscountedProductCount> possibleDiscounts, Basket basket)
         {
             // Looks for conflicting deals, removes them and favors best value deals
-
             List<Tuple<int, int, string>> conflictingIndexes = new List<Tuple<int, int, string>>();
             Dictionary<int, List<int>> handleDuplications = new();
             List<DiscountedProductCount> possibleCombinations = new List<DiscountedProductCount>();
@@ -30,7 +28,7 @@ namespace exercise.main.Discount
             for (int i = 0; i < possibleDiscounts.Count; i++)
             {
                 bool noConflict = true;
-                //for (int j = i + 1; j < possibleDiscounts.Count; j++)
+                // for each discount, check if there exist another discount for the same products, filter out
                 for (int j = 0; j < possibleDiscounts.Count; j++)
                 {
                     if (j == i)
@@ -40,12 +38,13 @@ namespace exercise.main.Discount
 
                         if (possibleDiscounts[i].SKU_amount.ContainsKey(key))
                         {
+                            // adds a list if dictionary doesn't have one for current index
                             if (!handleDuplications.ContainsKey(i))
                                 handleDuplications[i] = new();
                             if (!handleDuplications.ContainsKey(j))
                                 handleDuplications[j] = new();
 
-                            //if (!handleDuplications[i].Contains(j))
+                            // Avoid adding the same deal twice...
                             if (!handleDuplications[j].Contains(i))
                             {
                                 conflictingIndexes.Add(new(i, j, key));
@@ -61,8 +60,11 @@ namespace exercise.main.Discount
                 }
             }
 
+            // Add the non-conflicting possibleCombinations to bestDeal
             var bestDeals = new List<DiscountedProductCount>(possibleCombinations);
             float total = basket.getUndiscountedTotal();
+
+            // Compare the conflicting deals, select the most valuable
             foreach (var c in conflictingIndexes)
             {
                 int index_i = c.Item1;
@@ -78,13 +80,11 @@ namespace exercise.main.Discount
                     bestDeals.Add(possibleDiscounts[index_j]);
 
                 }
-
             }
 
             return bestDeals;
         }
 
-        //public Order calculateDiscount(Basket basket)
         private List<DiscountedProductCount> getPossibleDiscounts(Basket basket)
         {
             List<DiscountedProductCount> possibleDiscounts = new List<DiscountedProductCount>();
@@ -107,9 +107,70 @@ namespace exercise.main.Discount
             var productList = basket.getProducts();
 
             // Remove Discounted products from the product list...
+            productList = filterAwayDiscountedProducts(bestDealsDiscounts, productList);
+
+            Dictionary<string, OrderData> nameAmountPrice = new Dictionary<string, OrderData>();
+
+            createProductDataDictionary(bestDealsDiscounts, nameAmountPrice);
+            Dictionary<string, int> amontPerSku = GetAmountPerSkuDictionary(productList);
+
+            foreach (var product in amontPerSku)
+            {
+                float defPrice = _inventory.getPrice(product.Key);
+
+                nameAmountPrice[product.Key] = new OrderData
+                {
+                    name = product.Key,
+                    amount = product.Value,
+                    //individual_price = defPrice,
+                    discounted_price = 0.0f, // TODO: fix
+                    total_price = defPrice * product.Value,
+                    saving = 0.0f,
+                };
+            }
+
+            return nameAmountPrice;
+        }
+
+        private static Dictionary<string, int> GetAmountPerSkuDictionary(List<Product> productList)
+        {
+            Dictionary<string, int> amontPerSku = new Dictionary<string, int>();
+            foreach (var product in productList)
+            {
+                if (!amontPerSku.ContainsKey(product.SKU))
+                {
+                    amontPerSku[product.SKU] = 0;
+                }
+                amontPerSku[product.SKU]++;
+            }
+
+            return amontPerSku;
+        }
+
+        private void createProductDataDictionary(List<DiscountedProductCount> bestDealsDiscounts, Dictionary<string, OrderData> nameAmountPrice)
+        {
+            foreach (var product in bestDealsDiscounts)
+            {
+                string nameStr = "deal: " + string.Join(" + ", product.SKU_amount.Keys.Select(x => _inventory.getName(x)));
+
+                nameAmountPrice[nameStr] = new OrderData
+                {
+                    name = nameStr,
+                    amount = product.discountMultiple,
+                    //individual_price = 0.0f, // TODO: fix ...
+                    discounted_price = product.finalPrice,
+                    total_price = product.finalPrice,
+                    saving = product.possibleSavings,
+                    UsedDiscount = product.discount
+
+                };
+            }
+        }
+
+        private static List<Product> filterAwayDiscountedProducts(List<DiscountedProductCount> bestDealsDiscounts, List<Product> productList)
+        {
             foreach (var discount in bestDealsDiscounts)
             {
-
                 foreach (var di in discount.SKU_amount)
                 {
                     var sku = di.Key;
@@ -129,70 +190,10 @@ namespace exercise.main.Discount
                             counted++;
                         }
                     }
-
                     productList = temp;
-
-
                 }
             }
-
-            //Dictionary<string, Tuple<int, float>> nameAmountPrice = new Dictionary<string, Tuple<int, float>>();
-            Dictionary<string, OrderData> nameAmountPrice = new Dictionary<string, OrderData>();
-
-            foreach (var product in bestDealsDiscounts)
-            {
-                //string nameStr = "deal: " + string.Join(" + ", product.SKU_amount.Keys);
-                string nameStr = "deal: " + string.Join(" + ", product.SKU_amount.Keys.Select(x => _inventory.getName(x)));
-
-                nameAmountPrice[nameStr] = new OrderData
-                {
-                    name = nameStr,
-                    amount = product.discountMultiple,
-                    individual_price = 0.0f, // TODO: fix ...
-                    discounted_price = product.finalPrice,
-                    total_price = product.finalPrice,
-                    saving = product.possibleSavings,
-                    UsedDiscount = product.discount
-
-                };
-            }
-
-            Dictionary<string, int> amontPerSku = new Dictionary<string, int>();
-            foreach (var product in productList)
-            {
-                if (!amontPerSku.ContainsKey(product.SKU))
-                {
-                    amontPerSku[product.SKU] = 0;
-                }
-                amontPerSku[product.SKU]++;
-            }
-
-            //var amontPerSku = basket.getAmountPerSku();
-            foreach (var product in amontPerSku)
-            {
-
-                //float? defPrice = basket.getDefaultPrice(product.Key);
-                //float? defPrice = this._inventory.getPrice(product.Key);
-                //if (defPrice == null)
-                //Debug.Assert(false, "defPrice cant be zero...");
-
-                float defPrice = _inventory.getPrice(product.Key);
-
-                //nameAmountPrice[product.Key] = new (product.Value, defPrice.Value);
-                nameAmountPrice[product.Key] = new OrderData
-                {
-                    name = product.Key,
-                    amount = product.Value,
-                    //individual_price = defPrice.Value,
-                    individual_price = defPrice,
-                    discounted_price = 0.0f, // TODO: fix
-                    //total_price = defPrice.Value * product.Value,
-                    total_price = defPrice * product.Value,
-                    saving = 0.0f,
-                };
-            }
-
-            return nameAmountPrice;
+            return productList;
         }
 
         public string stringify()
