+++ exercise.main/StoreFrontExecutor.cs
@@ -1,0 +1,18 @@
+﻿namespace exercise.main
+{
+    public class StoreFrontExecutor
+    {
+        IStoreFront storeFront;
+        public StoreFrontExecutor(IStoreFront storeFront)
+        {
+            this.storeFront = storeFront;
+        }
+
+        public void run()
+        {
+            storeFront.run();
+        }
+
+    }
+
+}
